Acceptance tests buildspec
This is the buildspec part of introducing automated acceptance tests,
run through CodeBuild.

In addition to the buildspec file itself, there are some small changes
to the acceptance tests themselves (using a different variable name to
conform with what I've used in the Makefile), and to the Makefile, to
allow variables to support using unique stack names/prefixes for each
CodeBuild job.