Generate pinned requirements.txt using pip-tools
This change renames existing requirements.txt to requirements.in, and
introduces a Make-driven process to generate requirements.txt with all
sub-dependencies pinned to a specific version.

The intention is that developers maintain loose version definitions in
requirements.in, then generate a version-pinned requirements.txt from
that as necessary.

The Makefile has a basic formula to generate pinned versions from a
requirements.in, however pip-compile can be used to more selectively
update existing requirement.txt files.

A phony Makefile target, 'generate-pip-requirements', generates a file
for all 'requirements.in' files found in the working tree that don't
already have one (or have an outdated requirements.txt file).