Include base image in artefacts (#223)
* [#222] Include base image in artefacts

This commit attempts to address #222 by moving the Docker base image
retrieval to the build/release CI stage and including it in the release
artefacts. This version is then loaded in CodeBuild using `docker load`.

Although bundling the docker base image in the packaged artefacts isn't
particularly elegant, it addresses some constraints we have at this
particular time around creating new distribution channels. We may be
able to revisit and improve upon this later.

* Improve packaging targets

A few things in this:
 - Limit 'package' to not include all of backend/, just the lambda
   functions + layers
- Make exclusion patterns common between package/package-artefacts

* v0.18 release