Add retries for S3 bulk object deletion calls (#306)
boto3 does not appear to retry when individual objects encounter
transient errors during a bulk deletion. This can lead to some jobs
ending in a partial failure due to the old objects returning an error.

This adds a simple retry mechanism where the delete_objects call will be
retried whenever there are any errors returned, with an exponential
backoff, with a maximum number of 10 attempts. Because the API will
return success for deleting an object that is already deleted, there is
no danger to re-attempting deletions that have already succeeded.