Explicitly turn off use of S3 ACLs for Web UI Bucket (#360)
This is needed due to default behaviour changes in S3[1]. Following the
changes, new S3 buckets will default to having ACLs disabled. When ACLs
are disabled, PUT operations specifying an ACL are rejected. This will
result in the Web UI failing to upload to the S3 bucket in new
deployments.

There are two options to solve this: explicitly enable ACLs, or to stop
using ACLs. I have chosen the latter. This means that access to the Web
UI bucket is now controlled by the bucket policy alone.

References:
1. https://aws.amazon.com/blogs/aws/heads-up-amazon-s3-security-changes-are-coming-in-april-of-2023/