Improvements for container image build process
This change:
* Removes bundling the container base image with a release
* Uses the public ECR repository to retrieve the base image
* Specifies a label to pin the base image to Debian 11.6 "bullseye"
* Adds a build step to upgrade Debian packages

Since bundling the base image was added in dd7438c, ECR Public has
become an option to reliably retrieve the base image. This saves us from
having to bundle it in the release artefacts and for deployments to be
pinned to the version at-release-time.

This, in combination with adding a package upgrade step the Dockerfile,
makes it possible to rebuild the container image to update system
libraries when security updates become available from the upstream image
publisher/operating system vendor.

To mitigate compatibility concerns, I have pinned the image to the
Debian 11.6 "bullseye" stable release. This should mean we get security
patches automatically, but not breaking version changes.