Update CloudFront to use OAC instead of OAI
This change updates the Web UI CloudFront distribution to use Origin
Access Control (OAC) for authentication to the S3 origin, instead of the
previous implementation using Origin Access Identity (OAI).

OAC was introduced in August 2022 to supersede OAI. The main benefit of
this change in S3F2 is that OAI is not supported in AWS regions launched
after December 2022, so this change will be necessary to add region
support in the future.