Add the deployment region to OAC Name property
This fixes an issue introduced in
995f9d90bc3049e8a77490bc15fa23d1b785b900 that affected users who have
multiple S3F2 instances deployed within the same AWS account, using the
same ResourcePrefix parameter, but in different regions.

The issue prevented the stack from deploying because CloudFront
resources are global (not namespaced to a region), and the name must be
unique within an AWS account.

Note, for multiple deployments within an account we'd recommend using
a different ResourcePrefix, but we are fixing this as it worked in
previous releases.